Making so that each pane in TS exports a CJS module
