--- dashboard/basicPreferences.js
@@ -1,1 +1,0 @@
-module.exports = require('./basicPreferences.ts').default
