--- pad/padPane.js
@@ -1,1 +1,0 @@
-module.exports = require('./padPane.ts').default
