--- profile/editProfilePane.js
@@ -1,1 +1,0 @@
-module.exports = require('./editProfilePane.ts').default
