--- scratchpad/paneWrapper.js
@@ -1,1 +1,0 @@
-module.exports = require('./paneWrapper.ts').default
