--- trustedApplications/trustedApplicationsPane.js
@@ -1,1 +1,0 @@
-module.exports = require('./trustedApplicationsPane.ts').default
